[develop] Change type from uint32_t to uint8_t for better suitability

+++ include/Ros/RosNodeManager.hpp
@@ -114,7 +114,7 @@ public:
    * @param argv The argument vector
    * @param domain_id ROS domain ID to use for the node (0-232)
    */
-  void initialize(int argc, char** argv, uint32_t domain_id = 0);
+  void initialize(int argc, char** argv, uint8_t domain_id = 0);
 
   /**
    * @brief Restarts ROS node with a new domain ID
@@ -123,7 +123,7 @@ public:
    * @param argv The argument vector
    * @return True if restart was successful, false otherwise
    */
-  bool restartWithDomain(uint32_t new_domain_id, int argc, char** argv);
+  bool restartWithDomain(uint8_t new_domain_id, int argc, char** argv);
 
   /**
    * @brief Shuts down the ROS node manager
@@ -158,7 +158,7 @@ private:
 
   std::atomic<bool> is_running_; /**< Indicates if the ROS node is running */
   bool use_namespace_discovery_; /**< Indicates if namespace discovery is enabled */
-  uint32_t current_domain_id_;   /**< Current ROS domain ID */
+  uint8_t current_domain_id_;    /**< Current ROS domain ID */
 };
 } // namespace manager
 } // namespace ros
